hot fix

+++ src/services/autor.service.js
@@ -1,4 +1,6 @@
 import autorRepository from '../repositories/autor.repository.js'
+import livroService from '../services/livro.service.js'
+import { ErrorHandler } from '../util/error.handler.js'
 
 async function createAutor (autor) {
   return await autorRepository.createAutor(autor)
@@ -13,6 +15,12 @@ async function updateAutor (autor) {
 }
 
 async function deleteAutor (id) {
+  const livros = await livroService.getLivrosByAutorId(id)
+
+  if (livros.length > 0) {
+    throw new ErrorHandler(403, `Autor ${id} possui livros cadastrados`)
+  }
+
   return await autorRepository.deleteAutor(id)
 }
